Removed HttpClientHandler dead code and folded .Core file.
