--- src/libraries/System.Net.Http/src/System/Net/Http/HttpClientHandler.Core.cs
@@ -1,66 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Globalization;
-
-namespace System.Net.Http
-{
-    public partial class HttpClientHandler : HttpMessageHandler
-    {
-        // This partial implementation contains members common to Windows and Unix running on .NET Core.
-
-        private bool _operationStarted = false;
-        private volatile bool _disposed = false;
-
-        public long MaxRequestContentBufferSize
-        {
-            // This property is not supported. In the .NET Framework it was only used when the handler needed to
-            // automatically buffer the request content. That only happened if neither 'Content-Length' nor
-            // 'Transfer-Encoding: chunked' request headers were specified. So, the handler thus needed to buffer
-            // in the request content to determine its length and then would choose 'Content-Length' semantics when
-            // POST'ing. In .NET Core, the handler will resolve the ambiguity by always choosing
-            // 'Transfer-Encoding: chunked'. The handler will never automatically buffer in the request content.
-            get
-            {
-                return 0; // Returning zero is appropriate since in .NET Framework it means no limit.
-            }
-
-            set
-            {
-                if (value < 0)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(value));
-                }
-
-                if (value > HttpContent.MaxBufferSize)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(value), value,
-                        SR.Format(CultureInfo.InvariantCulture, SR.net_http_content_buffersize_limit,
-                        HttpContent.MaxBufferSize));
-                }
-
-                CheckDisposedOrStarted();
-
-                // No-op on property setter.
-            }
-        }
-
-        private void CheckDisposed()
-        {
-            if (_disposed)
-            {
-                throw new ObjectDisposedException(GetType().ToString());
-            }
-        }
-
-        private void CheckDisposedOrStarted()
-        {
-            CheckDisposed();
-            if (_operationStarted)
-            {
-                throw new InvalidOperationException(SR.net_http_operation_started);
-            }
-        }
-    }
-}
